pyproject.toml
[tool.poetry]
name = "ls-py-handler"
version = "0.1.0"
description = "A simple FastAPI server with run endpoints"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ls_py_handler"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
uuid = "^1.30"
alembic = "^1.15.1"
sqlalchemy = "^2.0.39"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
psycopg2-binary = "^2.9.10"
aiobotocore = "^2.21.1"
pydantic-settings = "^2.8.1"
orjson = "^3.10.15"
opentelemetry-sdk = "^1.36.0"
opentelemetry-exporter-otlp = "^1.36.0"
opentelemetry-instrumentation-fastapi = "^0.57b0"
aiobotocore-otel = "^1.0.0.dev3"
opentelemetry-distro = {extras = ["otlp"], version = "^0.57b0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
ruff = "^0.2.2"
pytest-benchmark = "^5.1.0"
pytest-memray = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules, pycodestyle (`E`/`W`), isort (`I`), and more
select = ["E", "F", "B", "I", "W"]
ignore = ["B008"]
# Avoid auto-fixing these errors
unfixable = []
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".ruff_cache",
    "__pypackages__",
    "dist",
    "venv",
    ".venv",
    "build",
    "node_modules",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
